Imp: refactor generic VNFM
